Improved reactionmodule
