+++ DiscordBot.Modules/Utils/ReactionBase/ReactionModuleBase.cs
@@ -1,42 +1,35 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Discord;
-using Discord.WebSocket;
 
 namespace DiscordBot.Modules.Utils.ReactionBase
 {
-    public class ReactionContext
-    {
-        private readonly Cacheable<IUserMessage, ulong> _cachedUserMessage;
-        private readonly ISocketMessageChannel _messageChannel;
-        private readonly SocketReaction _reaction;
-
-        public ReactionContext(Cacheable<IUserMessage, ulong> cachedUserMessage, ISocketMessageChannel messageChannel, SocketReaction reaction)
-        {
-            _cachedUserMessage = cachedUserMessage;
-            _messageChannel = messageChannel;
-            _reaction = reaction;
-        }
-
-        public IUserMessage UserMessage => _cachedUserMessage.HasValue ? _cachedUserMessage.Value : null;
-        public SocketReaction Reaction => _reaction;
-        public ISocketMessageChannel MessageChannel => _messageChannel;
-    }
-
-    public class ReactionAttribute : Attribute
-    {
-        public string[] ReactionName { get; }
-
-        public ReactionAttribute(params string[] reactionName)
-        {
-            ReactionName = reactionName;
-        }
-    }
-
     public abstract class ReactionModuleBase
     {
         public ReactionContext Context { get; set; }
 
         public abstract Task<bool> ExecuteAsync();
+
+
+        /// <summary>Sends a message to the source channel.</summary>
+        /// <param name="message">
+        /// Contents of the message; optional only if <paramref name="embed" /> is specified.
+        /// </param>
+        /// <param name="isTTS">Specifies if Discord should read this <paramref name="message" /> aloud using text-to-speech.</param>
+        /// <param name="embed">An embed to be displayed alongside the <paramref name="message" />.</param>
+        protected virtual async Task<IUserMessage> ReplyAsync(
+            string message = null,
+            bool isTTS = false,
+            Embed embed = null,
+            RequestOptions options = null)
+        {
+            return await this.Context.Channel.SendMessageAsync(message, isTTS, embed, options).ConfigureAwait(false);
+        }
+
+        //    var msg = cachedMessage.HasValue ? cachedMessage.Value : await socketMessageChannel.GetMessageAsync(cachedMessage.Id);
+
+        public async Task<IMessage> GetMessageAsync()
+        {
+            return Context.Message ?? await Context.Channel.GetMessageAsync(Context.MessageId);
+        }
     }
 }
